fraud-detection-new-isoscape/requirements.txt
functions-framework==3.3.0
google-api-core
google-events==0.4.0
protobuf==4.22.1
cloudevents==1.9.0
google-cloud-firestore
earthengine-api==0.1.358
firebase-admin
scipy
pandas
numpy
google-auth